AUI-1415 All form fields are showing up as required
The asterisk indicating a field was required was showing up on all fields, even the optional ones. This is because of the changes on https://github.com/liferay/alloy-ui/commit/9cd24ee65a54a9a32c176f16027ef8d1ad7a7856, which was checking the required value as if it was an array/String/NodeList when it's actually a boolean.
While fixing this I noticed that the same problem was happening for the showLabel attribute, so I fixed it in the same way.